requirements.txt
pandas
openai
PyGithub
requests
pydantic>=2.0.0
pyyaml
python-dotenv
